Work around Chrome bug causing sidebar to become invisible
Work around a Chrome bug [1] that can cause the sidebar to become
invisible if:

 1. The sidebar app is loaded from a Chrome extension AND
 2. The current tab was opened by clicking a link inside the sidebar
    app in a different tab.

When the issue occurs, the sidebar web app loads and runs normally but
is just not visible on screen. Temporarily hiding and showing the iframe
resolves the problem. At the point when the temporary hide & show
occurs, the iframe is positioned to the right of the viewport, so the
user does not see the visibility change.

Fixes #516

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=753314